oil_change_tracker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
pydantic==2.8.2
python-dotenv==1.0.1
Jinja2==3.1.4
alembic==1.13.2
httpx==0.27.0
watchfiles==0.21.0
